refactor: split androidApp module, migrate to AGP 9 KMP plugin
- composeApp: pure KMP library via com.android.kotlin.multiplatform.library
- androidApp: plain AGP 9 application shell (built-in Kotlin) with manifest,
  MainActivity, resources, signing configs and versionCode
- drop deprecated kotlin.mpp.androidGradlePluginCompatibility.nowarn and
  AGP legacy-DSL bypass properties
- expose compose runtime/ui/foundation/material3 and voyager as api from
  the shared module; rename APKs via copy tasks (AGP 9 removed the
  applicationVariants API)
