+++ androidApp/src/main/assets/adi-registration.properties
