+++ androidApp/src/main/kotlin/com/seiko/keystoreviewer/MainActivity.kt
