+++ androidApp/src/main/res/mipmap-xxxhdpi/ic_logo.png
