+++ gradle.properties
@@ -7,9 +7,4 @@ org.gradle.jvmargs=-Xmx2048M -Dfile.encoding=UTF-8 -Dkotlin.daemon.jvm.options\=
 android.nonTransitiveRClass=true
 android.useAndroidX=true
 
-# KMP
-kotlin.mpp.androidGradlePluginCompatibility.nowarn=true
 
-# AGP 9: keep legacy KGP (org.jetbrains.kotlin.multiplatform) compatibility
-android.builtInKotlin=false
-android.newDsl=false
