+++ settings.gradle.kts
@@ -31,3 +31,4 @@ dependencyResolutionManagement {
 }
 
 include(":composeApp")
+include(":androidApp")
